requirements.txt
anyio>=4.7.0
jinja2>=3.1.2
mcp>=1.17.0
puremagic>=1.20
pyglove>=0.5.0.dev202510170226
requests>=2.31.0

# extras:vertexai
google-auth>=2.16.0

# extras:mime-pil
pillow>=10.0.0

# extras:ui
termcolor==1.1.0
tqdm>=4.64.1
